# Conflicts:
#	src/main/java/uk/gov/hmcts/opal/controllers/DefendantAccountApiController.java
#	src/main/resources/openapi/CommonStrict.yaml
#	src/main/resources/openapi/DefendantAccount.yaml

+++ src/integrationTest/java/uk/gov/hmcts/opal/controllers/LegacyDefendantAccountConsolidatedAccountsIntegrationTest.java
@@ -16,6 +16,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -166,7 +167,7 @@ class LegacyDefendantAccountConsolidatedAccountsIntegrationTest extends Abstract
         )).thenReturn(new GatewayService.Response<>(
             HttpStatus.OK,
             LegacyGetDefendantAccountConsolidatedAccountsResponse.builder()
-                .version(2L)
+                .version(BigInteger.valueOf(2L))
                 .consolidatedAccounts(List.of())
                 .build(),
             null,
@@ -286,7 +287,7 @@ class LegacyDefendantAccountConsolidatedAccountsIntegrationTest extends Abstract
 
     private LegacyGetDefendantAccountConsolidatedAccountsResponse legacyResponse() {
         return LegacyGetDefendantAccountConsolidatedAccountsResponse.builder()
-            .version(9L)
+            .version(BigInteger.valueOf(9L))
             .consolidatedAccounts(List.of(
                 legacyAccount(233302L, "233302C"),
                 legacyAccount(233301L, "233301C")
